# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_anthropic_stc_backend_internal_models.CreateNotificationRequest:
    properties:
      content:
        type: string
      title:
        type: string
      type:
        enum:
        - info
        - success
        - warning
        - error
        type: string
      user_id:
        type: integer
    required:
    - title
    - user_id
    type: object
...
  github_com_anthropic_stc_backend_internal_models.PaginatedResponse:
    properties:
      items: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_anthropic_stc_backend_internal_models.PermissionCategoryResponse:
    properties:
      category:
        type: string
      permissions:
        items:
          $ref: '#/definitions/github_com_anthropic_stc_backend_internal_models.PermissionResponse'
        type: array
    type: object
  github_com_anthropic_stc_backend_internal_models.PermissionResponse:
    properties:
      action:
        type: string
      category:
        type: string
      code:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      resource:
        type: string
      scope:
        type: string
    type: object
...
  github_com_anthropic_stc_backend_internal_services.ActivityStats:
    properties:
      by_action:
        additionalProperties:
          type: integer
        type: object
      by_resource_type:
        additionalProperties:
          type: integer
        type: object
      by_user:
        items:
          $ref: '#/definitions/github_com_anthropic_stc_backend_internal_services.UserActivityCount'
        type: array
      today_count:
        type: integer
      total:
        type: integer
    type: object
...
  github_com_anthropic_stc_backend_internal_services.UserActivityCount:
    properties:
      count:
        type: integer
      user_id:
        type: integer
      username:
        type: string
    type: object
  github_com_anthropic_stc_backend_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: Prowl Range Backend API
  title: Prowl Range API
  version: 0.1.0
paths:
  /activities:
    get:
      description: 分页获取操作日志，支持按资源类型、用户和动作过滤
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 资源类型
        in: query
        name: resource_type
        type: string
      - description: 用户 ID
        in: query
        name: user_id
        type: integer
      - description: 动作类型
        in: query
        name: action
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_anthropic_stc_backend_internal_models.PaginatedResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取操作日志列表
      tags:
      - activities
  /activities/export:
    get:
      description: 导出操作日志为 CSV 文件，支持过滤
      parameters:
      - description: 资源类型
        in: query
        name: resource_type
        type: string
      - description: 用户 ID
        in: query
        name: user_id
        type: integer
      - description: 动作类型
        in: query
        name: action
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: OK
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 导出操作日志 CSV
      tags:
      - activities
  /activities/statistics:
    get:
      description: 获取操作日志的统计信息，包括总数、今日数量、按动作和资源类型分组
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_anthropic_stc_backend_internal_services.ActivityStats'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取操作日志统计
      tags:
      - activities
...
  /auth/oauth/callback:
    get:
      description: 处理 OAuth 提供商回调
      parameters:
      - description: 授权码
        in: query
        name: code
        required: true
        type: string
      - description: 状态参数
        in: query
        name: state
        required: true
        type: string
      responses:
        "302":
          description: Found
      summary: OAuth 回调
      tags:
      - auth
...
  /notifications:
    get:
      description: 查询当前用户的通知列表，支持分页和仅未读过滤
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - default: false
        description: 仅未读
        in: query
        name: unread_only
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_anthropic_stc_backend_internal_models.PaginatedResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 查询通知列表
      tags:
      - notification
    post:
      consumes:
      - application/json
      description: 创建通知并实时推送给目标用户（需 notification:create 权限）
      parameters:
      - description: 创建通知请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_anthropic_stc_backend_internal_models.CreateNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 创建通知
      tags:
      - notification
  /notifications/{id}/read:
    put:
      description: 标记指定通知为已读
      parameters:
      - description: 通知 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 标记通知已读
      tags:
      - notification
  /notifications/read-all:
    put:
      description: 将当前用户所有未读通知标记为已读
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 全部标记已读
      tags:
      - notification
  /notifications/unread-count:
    get:
      description: 返回当前用户的未读通知数量
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
            - properties:
                data:
                  type: integer
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取未读通知数量
      tags:
      - notification
  /permissions:
    get:
      description: 获取全部权限
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_anthropic_stc_backend_internal_models.PermissionResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取权限列表
      tags:
      - permissions
  /permissions/categories:
    get:
      description: 按分类分组返回全部权限
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_anthropic_stc_backend_internal_models.PermissionCategoryResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 按分类获取权限
      tags:
      - permissions
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"